+++ Public/Invoke-PSDuckDb.ps1
@@ -9,15 +9,17 @@ USING sum(Population);
     #>
     [CmdletBinding()]
     [Alias('psduckdb')]
-    param(        
-        [string]$Command
+    param(
+        [string]$FileName,
+        [string]$Command,
+        [switch]$json,
+        [switch]$csv
     )
-
-    $conn = [DuckDB.NET.Data.DuckDBConnection]::new("Data Source=:memory:")
-    $conn.Open()
-    $duckCommand = $conn.CreateCommand()
-
-    if (![string]::IsNullOrEmpty($Command)) {
+    
+    if (![string]::IsNullOrEmpty($Command)) {        
+        $conn = New-DuckDBConnection
+        $conn.Open()
+        $duckCommand = $conn.CreateCommand()
         $duckCommand.CommandText = $Command
 
         try {
@@ -37,51 +39,68 @@ USING sum(Population);
     }
     else {
         $ExitOn = @("exit", "quit", "bye")
+        
+        if ([string]::IsNullOrEmpty($FileName)) {
+            $FileName = ":memory:"
+        }
 
         Write-Host "Welcome to PSDuckDB! $(Get-Date)"
         Write-Host 'Connected to ' -NoNewline
-        Write-Host 'a transient in-memory database' -ForegroundColor Red
-
-        while ($true) {
-            $targetCommand = Read-Host "PSDuckDB"
+        if ($FileName -eq ":memory:") {
+            Write-Host 'an in-memory database' -ForegroundColor Red
+        }
+        else {
+            Write-Host $FileName -ForegroundColor Red
+        }
         
-            if ($targetCommand -in $ExitOn) {
-                break
-            }
+        try {
+            $conn = New-DuckDBConnection -Path $FileName
+            $conn.Open()
+            $duckCommand = $conn.CreateCommand()
 
-            try {
-                $duckCommand.CommandText = $targetCommand
-                $reader = $duckCommand.ExecuteReader()
+            while ($true) {
+                $targetCommand = Read-Host "PSDuckDB"
+        
+                if ($targetCommand -in $ExitOn) {
+                    break
+                }
+
+                try {
+                    $duckCommand.CommandText = $targetCommand
+                    $reader = $duckCommand.ExecuteReader()
             
-                Out-DuckData $reader | Format-Table
+                    $returnData = Out-DuckData $reader
+                    if ($json) {
+                        $returnData | ConvertTo-Json
+                    }
+                    elseif ($csv) {
+                        $returnData | ConvertTo-Csv
+                    } 
+                    else {
+                        $returnData
+                    }
+                }
+                catch {
+                    Write-Host $_.Exception.Message -ForegroundColor Red
+                }
             }
-            catch {
-                Write-Host $_.Exception.Message -ForegroundColor Red
+        
+        }
+        catch {
+            Write-Host $_.Exception.Message -ForegroundColor Red
+        }
+        finally {       
+            if ($null -ne $reader) {
+                $reader.Dispose()
+            }
+            
+            if ($null -ne $duckCommand) {
+                $duckCommand.Dispose()
+            }
+            
+            if ($null -ne $conn -and $conn.State -eq 'Open') {
+                $conn.Close()
             }
         }
-        
-        if ($null -ne $reader) {
-            $reader.Dispose()
-        }
-        
-        $duckCommand.Dispose()
-        $conn.Close()
-    }
-
-}
-
-function Out-DuckData {
-    param($reader)
-
-    while ($reader.read()) {
-        # Create a hashtable for the current row
-        $rowObject = [Ordered]@{}
-        For ($columnIndex = 0; $columnIndex -lt $reader.FieldCount; $columnIndex++ ) {
-            # Add field name and value as key-value pair
-            $rowObject[$reader.GetName($columnIndex)] = $reader.GetValue($columnIndex)        
-        }
-    
-        # Convert the hashtable to a custom object and add it to the array
-        [PSCustomObject]$rowObject
     }
 }