+++ examples-sql-method/02-read-csv.ps1
@@ -1,0 +1,9 @@
+Import-Module $psscriptroot\..\PSDuckDB.psd1 -Force
+
+#psduckdb -c "SELECT * FROM '$PSScriptRoot\..\data\csv\sample\sales1.csv';" | Format-Table
+$db = New-DuckDBConnection
+
+$db.sql("SELECT * FROM '$PSScriptRoot\..\data\csv\sample\sales1.csv';") | Format-Table
+
+$db.CloseDB()
+