+++ examples-sql-method/05-read-wild-card-json.ps1
@@ -1,0 +1,9 @@
+Import-Module $psscriptroot\..\PSDuckDB.psd1 -Force
+
+# psduckdb -c "SELECT * FROM '$PSScriptRoot\..\data\json\sales*.json';" | Format-Table
+
+$db = New-DuckDBConnection
+
+$db.sql("SELECT * FROM '$PSScriptRoot\..\data\json\sales*.json';") | Format-Table
+
+$db.CloseDB()