+++ examples-sql-method/08-join-csv-files.ps1
@@ -1,0 +1,12 @@
+$db = New-DuckDBConnection
+
+# join customers and orders on customerid
+$sql = @"
+SELECT * 
+FROM '$PSScriptRoot\..\data\orders\customers.csv' c
+JOIN '$PSScriptRoot\..\data\orders\orders.csv' o
+ON (c.customerid = o.customerid);
+"@
+
+$db.sql($sql)[0]
+$db.closedb()